비 로그인시 예외처리

+++ portpolio-master/src/main/java/com/younghun/klom/controller/user/crud/UserDeleteController.java
@@ -20,9 +20,13 @@ public class UserDeleteController {
 	
 	// home으로 이동  // 유저 삭제 
 	@RequestMapping(method = RequestMethod.POST)
-	public String kill(HttpSession session, @ModelAttribute UserVo userVo) {
+	public String kill(HttpSession httpSession, @ModelAttribute UserVo userVo) {
+		if (httpSession.getAttribute("data") == null) {
+			return "/error/400";
+		}
+		
 		userService.delete(userVo);
-		session.invalidate();
+		httpSession.invalidate();
 		return "redirect:/";
 	}
 }
